CORE-18961 Reject fiber cache sandbox mismatches (#5511)
Rejecting the Fiber if there is a mismatch with sandbox (like it's done with suspend count), and additionally evicting the entry from the cache when this is discovered. This solves the issue when Fiber is pulled from the singleton Fiber cache after the sandbox in which the Fiber classes are instantiated was evicted from the cache. 

---------

Co-authored-by: Simon Johnson <simon.johnson@r3.com>