CORE-19350 - Disable vnode labels for p2p metrics (#5552)
Removing the labels for source & destination vnode at the source (worker). Otherwise, when one is using dropLabels in the helm charts, these are filtered out but when that leads to multiple timeseries that are the same for the same time range one of them is kept and the rest dropped. In this case, this was leading to under-reporting of message data by a few orders of magnitude in cases where there were a lot of virtual nodes with roughly the same traffic.