CORE-19241: Add a batch update API to the State Manager (#5519)
Problem statement
In many cases when batch processing events, multiple different state operation types must be performed as part of processing the same batch. For example, processing a group of flow events may result in some flows being created, some updated, and some deleted. In these cases, the state manager does not provide an API that allows all of these operations to be logically grouped together. This can have performance implications - for example, the state manager must create three transactions at present to perform all of the required operations.

Solution
Create an API that logically groups a set of state manager operations.

The implementation provided for the postgres backed state manager uses a single transaction to perform all operations (including a get to retrieve failed states). This should improve the performance of the event mediator when integrated, as it should cut a large number of database transactions out of processing (potentially up to two thirds of the transactions per processing loop).