CORE-19404: Change update vnode connection strings to nested objects (#5582)
This change is only making changes to update DB connectivity parameters as specified by the Jira. These were introduced in C5.2 and are relatively easy to change.

Similar sort of DB connectivity parameters are passed at vNode creation time. But this is a bigger change, which should be done with REST compatibility considerations in mind and may eventually be done as a separate piece of work.