CORE-19492 Adding subscription to read from FlowStatus cleanup topic and delete stale FlowStatuses (#5595)
This PR creates a new processor, `FlowStatusDeletionExecutor` which is intended to read `ExecuteFlowStatusCleanup` records from the `rest.flow.status.cleanup` topic.

These records contain the fingerprint of a `FlowStatus` record in the `StateManager` (I.e., key and version). For each `FlowStatusRecord` in the `ExecuteFlowStatusCleanup`, we create a dummy `State` object; all of these objects are then passed to the `StateManager::delete()` function to enact a bulk delete.

We retain the version from the original deletion request to preserve optimistic locking in the `StateManager`. If any `FlowStatus` record is updated between the time that it's picked up by the scheduled job and the time that we attempt deletion, deletion for that record will fail. This will be logged, but execution will proceed with no errors.