CORE-19461: state for individual rotations complete (CryptoRewrapBusProcessor) (#5565)
This PR handles the state manager updates of managed key rotation. The testing used by the class also had to change how its serializers were generated, as both the strategy used previously and the strategy used in #5539 did not work. The previous strategy didn't work with multiple serializer types being returned, as mockito did not allow us to change based on generics, and the strategy used in the linked PR did not allow for the return of multiple deserialisers, so this PR had to switch to having multiple factories, one for managed tests and another for unmanaged tests.