CORE-19572 Robust message processing (#5613)
In the JIRA ticket we see bad permissions set up for a vnode crypto db. The code assumed all tenantIds had accessible crypto dbs and threw an exception if not. In the case of bad user permissions (possible for example if bringing your own vnode db) this is not the case, so we need to be more robust.

Catch exceptions and ignore where possible to enable as much of key rotation as possible to continue.