CORE-19343 Respond back for sessions that cannot be decrypted (#5601)
In the scenario where a link manager receives an inbound message for a session that exists but cannot be decrypted, a message is sent back to the source notifying that the session has been lost, so that a new session can be eagerly established, and message replays for that session can be stopped. This message is sent securely via an outbound session from the destination cluster.

Note: Enables shared link manager sessions.