CORE-19592: vNode wrapping key rotation never reports as finished (#5655)
When running vNodes wrapping key rotation, it never reported back as finished. The problem was that we stored the record in the state manager using key generated by wrapping key's alias. This alias was never available when we tried to update the record (in RewrapBusProcessor) so we tried to find the record using wrapping key's uuid, which could never worked.

Change:
- Use key's uuid opposed to alias as state manager record's key as it's available in both crypto processors dealing with state manager records.