CORE-19444 UTXO ledger batch inserts (#5627)
Batch inserts in the UTXO ledger persistence code.

Postgres-specific insert statements following the structure of `INSERT INTO table VALUES (?, ?, ?), (?, ?, ?)` where `(???)` is repeated multiple times so that multiple rows can be inserted within a single insert statement. Batching in this way significantly improves performance, especially compared to using the _generic_ form of batching that `PreparedStatement` provides.

All UTXO ledger inserts previously inserted by looping over collections and inserting a row one at a time have been switched to insert in this way.

`BatchPersistenceService` has been added to encapsulate the batching logic to a single class. This takes care of two levels of batching:
- Batching the number of rows inserted per insert statement.
- Batching the number of inserts included in a batch of statements before contacting the database.

To support this behaviour change, queries need to be dynamic to be told the number of rows included in a single statement and then produce a query with the correct number of `(???)` / value blocks.