CORE-18719 implement the session event processor (#5577)
Add the SessionEventProcessor which signals to other Session managers when session negotiation is complete and when a session is rotated. When this signal is received Session Managers will look in the statemanager for the session (this is done to avoid needing to worry about reconciliation).

I also changed session rotation, so that only cached sessions are rotated. To prevent too many cached sessions being stored in a hashmap.