CORE-19493 Adding FlowStatusCleanup job to SchedulerProcessorImpl (#5631)
This PR replaces the `FlowStatusCacheService` with the new `FlowStatusLookupService`, backed by out-of-process state storage. It enables a scheduled job for cleaning up stale flow statuses from storage after a default period of 7 days. As part of this change, the web socket functionality for retrieving flow status updates has been removed.