CORE-19303 - on sync rpc processing failures, set whatever checkpoint we have (#5660)
Problem description

Some flows are left in START_REQUESTED when external event processing fails. When no checkpoint has been written, we write a state with null checkpoint and the PROCESSING_FAILURE error flag in the metadata.

The cleanup logic in TimeoutEventCleanupProcessor ignores this state because the checkpoint is null. Therefore the flow is never cleaned up, no flow status to indicate the failure is published.

Solution

We actually have some state accumulated because the flow engine has processed an event, and produced an output. Store this partially updated state, which contains checkpoint. Flag it as PROCESSING_FAILURE so it is ignored and not fed back into the processing queue. The cleanup task can now clean this up as it has the checkpoint.