CORE-19563: Remove check that previous key rotation has finished (#5625)
This PR removes the check that a previous key rotation had finished in both unmanaged and managed key rotations. Additionally, it adds a check that the state manager is active, which would previously throw an error as a side effect of checking the previous rotation status. Technically, the check in `KeyRotationRestResource` will throw the error from trying to access `stateManager.isRunning`, but the code was written to throw the same error that is thrown by the failed attempt at accessing it.