CORE-19372 Use databases for cluster schema (#5643)
This PR removes the concept of a "cluster DB". Instead, the databases
stanza introduced for state managers can now also be used for the
config, crypto, RBAC, and virtual node databases. By default, they will
all share a database as before but they can be split apart. The other,
perhaps more significant, change, is that the automatic bootstrap will
now create separate users for each of crypto, RBAC, and virtual node
databases, and one for worker that accesses the config DB. It is,
therefore, now only necessary to give a single bootstrap user for each
database and everything else is handled by the chart.