CORE-19535 get the hash of the input parameters in a deterministic way (#5562)
This allows the external event framework to deterministically generate external event requestIds

A hash of the input object is required as part of the id generator. This is to defend against non deterministic client code, sending different outputs when a flow fiber is rerun.

For example
if (Instant.now().toEpochMilli() % 2 == 0L) { persist(dog) } else { persist(cat) }

If a flow ran and executed persist(dog), received a success response and then the mediator failed the flow will rerun. If the rerun resulted in persist(cat) and the id generator did not account for the payload, then entity persistence worker would detect a duplicate and return success even though dog was persisted, not cat.

In order to generate the hash of the inputs, it is now mandated that classes passed in as PARAMETERS be @CordaSerializable

This PR adds @CordaSerializable to the existing PARAMETERS objects passed into the external event request executor. It also refactors some of these classes to not pass in objects that are not serializable, such as ByteBuffers.