CORE-19467 Enable snappy compression (#5661)
Enable snappy compression but also provide a way to serialize without
compression for our internal code.

This is needed for signature verification where we need to ensure that
signature verification works when communicating with 5.1 nodes who
cannot serialize with compression themselves. Turning off compression
here means that all parties whether on 5.1 or 5.2 will generate the same
byte output, allowing signatures to be verified.