CORE-19503 - Updated the Helm charts (#5700)
The cli options `--mediator-replicas-flow-session-in`, `--mediator-replicas-flow-session-out`, and `--mediator-replicas-flow-session` have been previously added in. 

This commit is a follow up work to allow users to set the cli arguments from a Helm chart.  The field will be hidden from customers. The `value.yaml` file should look like the following when setting the new config options:
workers:
  flow:
    extraArgs:
      - "--mediator-replicas-flow-session=1"
  flowMapper:
    extraArgs:
      - "--mediator-replicas-flow-session-in=1"
      - "--mediator-replicas-flow-session-out=1"