CORE-19695 Handle duplicate signed and filtered transactions (#5709)
Update `findSignedTransaction` to not return signed transactions where a transaction record with `status = VERIFIED` and `is_filtered = true`. These represent filtered transactions to attempting to return a signed transaction is wrong.

Update `persistTransaction` to overwrite transaction records with `status = VERIFIED` and `is_filtered = true`. We need to be able to do this so verified signed transactions can overwrite the existing record. The `is_filtered` flag switches to false at this point.

Add tests to verify this logic amongst other scenarios.