CORE-19695 Handle duplicate signed and filtered transactions (#5728)
Support having unverified and filtered transactions existing for the same transaction simultaneously. In this case, the `status` would be `U` and `is_filtered = true`. The filtered transaction would be used until the transaction status switches to  `V`.

Fix HSQLDB queries that don't seem to use the `IN` SQL clause correctly. Switching to `=` and `OR` worked.