CORE-19738 Hash merkle proof id
Store the merkle proof id as a hash of
`<transactionId>;<groupIndex>;<leafIndexes>`. The same structure was
used before this change, but would easily go over the columns size limit
when storing filtered transactions.

Using the hash gives us predictable ids as it is based on computed data
rather than any randomisation and also keeps the ids shorter.

As the id contained the leaf indexes which were parsed when read
from the database, the content still needed to be preserved in a
readable format; a column has been added for `leaf_indexes` to preserve
this data.