CORE-19834 Update missing transaction in backchain resolution message
Update the error message and add a log line for when a signed
transaction is missing during the sending of a transaction backchain.
This can occur when a sender has received part of the backchain as a
filtered transaction and then tries to send the whole backchain. Upon
reaching the filtered transaction in the chain, an error is thrown due
to no signed transaction existing.

The error message makes it clearer why this could have happened to
users.