CORE-19285 - Fix bug caused by creating instead of updating state when replay happens (#5785)
Flows were stuck in RUNNING state, because participants were unable to establish sessions.
When the ResponderHello message is lost and an InitiatorHello got replayed by the initiator: the responder side started processing the message again. However, instead of eventually getting past through this step and sending the ResponderHello again, we got into a loop where the processing of the init message failed, because we tried to create new states using same ID, instead of updating the existing state from the first processing.