CORE-19319 - Modify load generation for baseline performance testing (#5771)
Currently, in runScenario.sh we send a given number of messages on each iteration with increasing the batch size only. We had to introduce a new parameter that defines a time window for load generation, instead of defining the total number of messages. Otherwise, as we increase the batch size the duration of each test is increasingly smaller.
Ideally, we should be running load for ~5 minutes before moving on to higher load.