CORE-14391 - Set an initial status for a `CPI UPLOAD` request (#5811)
Set an initial status for a `CPI UPLOAD` request to decrease the chance of returning `400` code while the request waits in the Kafka queue to be processed by the `db worker`.

Keep in mind this is only a mitigation. It is still possible that a 400 error code is returned when multiple REST workers are used. This happens because Corda follows the `eventual consistency` pattern.