CORE-14483 - Log entries generated inside Flow code now have a trace ID in the MDC (#5844)
Updated the code so the `traceId` is preserved when logs are generated by the flow. This helps fetching the relevant logs when troubleshooting bugs and reduces the risk of overwhelming Grafana.
Added the `traceContext` to `FlowFiberExecutionContext`.
Replaced the underscore by a dot so the mdc names are consistent in the logs.
The function `LogUtils.clearMDC()` should only clear the properties that are set explicitly by Corda. The MDC properties set by third-party dependencies should not be cleared out.
Renamed the MDC variables in `AuthorizationUtils` to follow the same naming pattern found in `LogUtils`
Moved the MDC variables in `AuthorizationUtils` to `LogUtils` to make it easier to find all MDC related variables.
Added the prefix `corda.` to the MDC related strings so we can tell which ones are set by Corda explicitly.