CORE-19985 Stop bringing coordinator `DOWN` on new config (#5847)
When the component goes `UP`, then on processing subsequent config change event we would take its coordinator to `DOWN` and `UP` again. This was causing the depending reconciler to go `UP` then `DOWN` then `UP` again. This seems to have been causing the issue of staring concurrent reconciliations (for the same data).