CORE-16941 OSGi refactor (#5830)
Non-functional tidying up. Formalised the class relationship in the OSGi wrapping framework.

Bundle observer is dispatched by Felix on it's own thread, the original implementation failed to accommodate async nature of this callback and so gave faulty logging. Logging was incorrectly warning when things were activated ok (there was a mismatch between our stored state of activation and the sync getter function from the Bundle). The bundle observer is now for information only for debugging purposes.

NO_TIMEOUT waiting for countdown latch offered nothing functionally at all so was removed.

Lazy Activation support. Not all bundles will be active when our OSGi wrapping framework activates them, this is because bundles can (and are) marked by the creators as lazy activating.

This change also activates OSGi info logging in order to track down the root cause of CORE-16941. It will spam the worker log on startup, so we should turn this back to warn only as soon as we have the necessary logs.