CORE-19896 timestamp check in query (#5888)
Introduce a storage type for timestamps, and store the last persisted timestamp in the flow checkpoint if it's later than the currently stored time. When running a query as of now, i.e. without providing a timestamp, use the later of now() and the stored timestamp so that if the persistence worker's clock is ahead (clock skew between machines), we still can query everything that is already persisted.

--
Co-authored-by: nkovacsx nandor.kovacs@r3.com