CORE-19313: Prepare TokenSelectionTests for new test
This PR removes the cluster health check because it is not a concern of this test. Corda 5 is a HA system; it should provide working REST endpoints and process work even if some workers are unhealthy.

It removes the test order annotations because the test does not depend on order.

The `beforeAll` method has been made static because JUnit `BeforeAll` methods are expected to be static and working around that convention by applying annotations is unnecessary work to fight established conventions.