CORE-20047 Invalidate the flow fiber cache when new partitions are assigned to clear out potentially stale fibers (#5905)
Problem
This PR solves a problem whereby stale Fibers from the FlowFiberCache can be read and used instead of a more recent version of the fiber.

Scenario
Flow starts on Pod A which is assigned flow.start topic. Flow reaches the point where it needs to receive from counterparty, puts fiber in cache and then tries to save the checkpoint. This will fail if the task exceeded its timeout. A rebalance occurs. Pod A is assigned the flow.session topic and Pod B is assigned the flow.start topic. Pod B replays the start flow event successfully and writes the state. Flow progresses to the same suspension point and initiates a session with a counterparty successfully.

Counterparty responds with the first message sent to the flow.session topic and is picked up by Pod A.
Pod A has the first vesion of the fiber in its cache which is invalid and reuses it. This then results in an error as the FlowSession ID does not match the message received.

Solution
Invalidate the cache on a rebalance to clear out potentially stale fibers in the cache.