CORE-20046 - cancel task manager threads after timeout (#5910)
Problem Description

When a task is timed out from mediator side via the CompletableFuture API, the task still exists and processes flows. If the mediator expects this thread to be stopped so it can re-run a batch, we run into a race between two threads processing the same flow and writing to the flow fiber cache.

Solution

Calling CompletableFuture.cancel(true) on the future returned from the CompletableFuture.supplyAsync API does not interrupt the thread. However, the future returned from directly submitting the task to the executor does interrupt and stop the task when calling cancel. The thread will only be interrupted when a blocking call happens, or when we check if the current thread is interrupted. Doing so in the FlowFiberCache's put() method will interrupt the thread if it tries to write to the cache when it should be interrupted. Another blocking operation is httpClient.send() which now propagates the InterruptedException back to the task error handling in the task manager.