CORE-20046 - flow fiber cache invalidate on get (#5925)
**Problem description**

A flow with the following code cannot be interrupted using `future.cancel()` as there is no blocking operation call to give opportunity to interrupt the thread:
```
        while (counter < 100000000000) {
            counter++
        }
```
If this flow has initiated a session to receive a message, and on receiving the message then hangs in this while loop, the thread will still be running the fiber while the retry thread comes along to re-process the received session event. Thus, the "Not new or suspended" error message.

**Solution**

Remove from flow fiber cache when getting from the cache. The retry thread will deserialize the checkpoint rather than using the cache, if the hanging thread later continues, it won't be able to overwrite the fiber as its thread has flag `isInterrupted=true` and we already have an interrupted validation check in the `FlowFiberCache.put()` method.