CORE-20186 Adding user-provided deduplication ID to `persist` APIs (#5920)
This PR changes the API of entity persistence in a breaking way. It's now required to pass a unique (within the context of a flow), deterministic deduplication ID with persist() (single or bulk) request to enable request idempotence in failure scenarios.
The maximum size of this new dedeplication ID is 128 chars, and it will be maintained plaintext in the system (meaning that it will be visible as-is in the logs to aid debugging).

The ExternalEventExecutor has been modified to try an implicit cast of the request parameters, PARAMETERS: Any, to PersistParameters. If this succeeds, we define our requestID as {flowId}-{deduplicationId}. Otherwise, we fallback to the original definition of {flowId}-{parameterHash}-{suspendCount}.