CORE-19220 Delivery tracker message cache (#5977)
The data message cache is responsible for storing data messages in memory while they are tracked. To avoid the cache growing too large, two constraints are put on the cache - the first is an overall size based on the total storage space represented by all messages in the cache, the second is an age of items in the cache based on Kafka offset.