CORE-20358 CORE-19896 backport handling of clock skew in queries (#6025)
This PR returns timestamps from the persistence worker to the finality flow since we need to know the timestamp from DB when transactions are stored to be able to debug the current timestamp skew between different pods.

removes a check whether the incoming timestamp is future time to prove timestamps in different queries have time skew.

Introduce a storage type for timestamps, and store the last persisted timestamp in the flow checkpoint if it's later than the currently stored time. When running a query as of now, i.e. without providing a timestamp, use the later of now() and the stored timestamp so that if the persistence worker's clock is ahead (clock skew between machines), we still can query everything that is already persisted.

--
Co-authored-by: nkovacsx nandor.kovacs@r3.com
Co-authored-by: Jenny Yang <135036209+jennyang-r3@users.noreply.github.com>