CORE-20405 Kryo serialization error in query object
`VaultNamedParameterizedQueryImpl` added a constructor parameter of
`FlowFiberExecutionContext` which lead to a kryo serialization error as
`flowFiberExecutionContext` should not be serialized.

This occurred due to a backport from 5.3, where a service that exists in
5.3 does not exist in 5.2.

This has been fixed by using `FlowFiberService` to get hold of the
correct objects inside a method that won't be kryo serialized.