CORE-20247: InvalidParamsException: unable to find master wrapping key XXXXX in tenant YYYYY (#6041)
In rare cases we have got a clash when generating key alias. This caused wrapping key not being saved in tenant's db as the alias was already cached as used. When generating key pair for the alias, wrapping key couldn't be find and the operation has failed.
The clashes happened, because we converted random uuid into string and then into bytes before converting it to hex and taking 12 chars. Conversion to string diluted the randomness of the uuid.
We are introducing a new function that converts uuid into bytes directly, keeping the randomness and therefore preventing the clashes to happen much less often.

Here is the output of random uuid (0), using new conversion + toHexString (1) and using old conversion toString().toByteArray() + toHexString (2)

0: 4b40ae93-2342-4295-9ee7-7b0e4dff12cb
1: 4B40AE93234242959EE77B0E4DFF12CB
2: 34623430616539332D323334322D343239352D396565372D376230653464666631326362
0: 2b57f797-ab8c-47ff-9dfa-3f9c2f4d836b
1: 2B57F797AB8C47FF9DFA3F9C2F4D836B
2: 32623537663739372D616238632D343766662D396466612D336639633266346438333662
0: 83452f7a-a4db-4b73-91ca-e8130cb5a0d7
1: 83452F7AA4DB4B7391CAE8130CB5A0D7
2: 38333435326637612D613464622D346237332D393163612D653831333063623561306437