CORE-20203: Add a verification exception categorizer and default errors to fatal (#6015)
Adds a verification processor exception handler and defaults errors to fatal.

Previously errors would default to transient. This leads to issues with the flow engine, as it retries transient errors indefinitely. If the error is not transient, this results in flows getting stuck.

Additionally, existing errors such as CPIs becoming missing are recategorized, which ensures that they are retried but not indefinitely.