CORE-20332 logic to understand and skip metadata json blob (#6046)
This PR is for 5.2.1 feature.
Made parseMetadata() to have a logic that there could be a header in a json blob of a metadata transaction.
Added logic to find metadata schema based on version returned from consume()
introduced 5.2.1 platform version

FYI: This PR doesn't implement header of JSONB. Implementation will be done in other ticket.