CORE-19951 Notary database errors (#6053)
Concurrent notarization of the same transaction can lead to constraint violations in the uniqueness database due to inserting to the transaction and state tables.

Consumption of the same state concurrently is already handled by the existing code and sets the basis for handling these errors in this change.

Errors when executing a database transaction in `JPABackingStoreImpl.executeTransaction` are retried (depending on the error). When a particular error occurs, the whole `executionTransaction` block is replayed in a new database transaction. This gives concurrency errors a chance to recover as the uniqueness checks will be re-executed, with the problematic inserts/updates being removed by these checks.

This behaviour has been leveraged for a larger set of database errors, not just constraint violations, which also puts its behaviour more in line with the ledger persistence code.

`PersistenceExceptionCategorizer` has been moved so it can be used from the ledger persistence and uniqueness code. The handling of errors has been further generalised to support the extra usage.