CORE-19840 Send backchain when receiver has filtered transaction (#6092)
Changed a number of methods to only return information about signed transactions, ensuring that filtered transactions were not leaking through in the returned data.

Doing this allowed a transaction to be sent by `sendTransaction` and `sendTransactionWithBackchain`, however, this highlighted an issue where storing a filtered transaction and then overwriting it with a signed transaction led to the transaction signatures having the wrong indexes. This caused duplicate and missing signatures (the primary issue).

To resolve the transaction signature issue, the signature table has been changed to make the `transaction_id` and `pub_key_hash` the primary key, and the `signature_idx` has been dropped altogether. This resolves the issue because a `transaction_id` and `pub_key_hash` pair will always be unique, so we can always correctly identify new signatures via the key. Conflicts will continue to be resolved by `ON CONFLICT DO NOTHING`.

Additional changes were made to change the `signatures` field in `UtxoSignedTransactionImpl` and `UtxoFilteredTransactionAndSignatures` to be a `Set` to further align with the changes made in the database.