CORE-20284 Added handling to CryptoFlowOpsProcessor for transient exceptions (#6098)
Adds an exception handler to the crypto service for categorizing exceptions that propagate back to the flow engine.

This uses the same logic as the persistence worker to identify exceptions that can be retried vs those that are fatal. It also changes the default behaviour to mark exceptions as platform (i.e. communicate back to the user code) rather than transient (retry indefinitely).

---------

Co-authored-by: James Higgs <james.higgs@r3.com>