CORE-9164 - Persist mgm info on member side (#6082)
To simplify the member list handling and reconciliation, we persist the MGMs member info to the member’s member list database table at the same time as publishing it to Kafka. To simplify the reconciliation process in case of Kafka failure.

Before this change, during virtual node creation and upgrade, we parse the MGM information from the group policy file in the CPI and publish it to kafka. This means that the information only exists in the member list on Kafka, or in the group policy file.
For other members (i.e. everyone other than the MGM) the member list will be received from the MGM, persisted to the database, and published to kafka.
In the case of Kafka failure, the virtual node database reconciliation will help recovery by reconciling the member list from the database to Kafka. Without this change, the MGM info would need to be retrieved from the group policy (making reconciliation unnecessarily complex).

Because the VirtualNodeInfoProcessor could not process the vnode record required to access the vnode vault DB in time, the MGM info publishing and persistence was moved to the dynamic registration service instead.