CORE-20434: Replace crypto cluster level tenant with cluster crypto db schema name (#6137)
Previously, we used `CryptoTenants.CRYPTO` as a placeholder to connect to the cluster database - crypto schema. Crypto tenant never owned the asymmetric keys.
It was very confusing as we have a proper crypto tenant P2P, who can have asymmetric keys and does all sort of key operations which never happened for crypto tenant.

This change is replacing crypto tenant with `ClusterCryptoDb.SCHEMA_NAME` which automatically suggests on higher levels that we want a connection to a cluster crypto database.