CORE-20675 Fix race condition in P2P lifecycle (#6189)
Prevent subscriptions starting before historic Avro Schema's are published, in the Link Manager and Gateway. These historic schemas are read from Kafka, so we must read them first before we use them.

The code inside AvroSchemaProcessor publishes a SetupConfigSubscription event to the ConfigReadServiceEventHandler which only publishes config after this event has happened. Hence if we wait for the config to be published before starting subscriptions the AvroSchemaProcessor must have published an initial snapshot of the schema data.